requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.2
elasticsearch==7.6.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
six==1.14.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8